Replace Rc with Gc to plug memory leaks
The implementation of closures in the previous commit introduced a
memory leak due to a reference cycle in which a LoxFunction would
capture an Rc to its enclosing environment, but its environment would
also already have an Rc to the function.

I couldn't see any simple way of breaking this cycle via weak refs or
other techniques, so I decided to bite the bullet and pull in Gc. I had
been planning to do this eventually anyway, since as a garbage collected
language, Rc was never going to be enough for Lox on its own.

The current implementation probably doesn't need to use Gc to the extent
it does, and could likely get away with Rc on some types that are
guaranteed to never participate in cycles. But at least for now we're
all in on Gc, mostly just to keep things simple.

Speaking of keeping things simple, this commit also eliminates a few
instances of trait objects from the codebase, mostly to avoid writing
custom Trace implementations for them. They've been replaced with enums,
and as such are probably more ergonomic and performant than the trait
objects anyway.